mcp-servers/image-gen-mcp/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "image-gen-mcp"
version = "1.0.0"
description = "Multi-provider image generation MCP server for AGI visual communication"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "AGI System", email = "agi@local" }
]
keywords = [
    "mcp",
    "image-generation",
    "ai",
    "stable-diffusion",
    "flux",
    "agi"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "mcp>=1.0.0",
    "aiohttp>=3.9.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
image-gen-mcp = "image_gen_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/image_gen_mcp"]
